Add features stories for breadcrumbs
